requirements.txt
tensorflow
opencv-python
django==4.0
djangorestframework
djangorestframework-simplejwt
numpy
pyttsx3
gTTS
Pillow